refactored queue logics
