+++ .env.example
@@ -2,4 +2,3 @@ PORT=3000
 MONGO_URI=mongodb://root:example@localhost:27017/mydb?authSource=admin
 REDIS_CONFIG_HOST=localhost
 REDIS_CONFIG_PORT=6379
-WORKER_COUNT=2
